requirements.txt
numpy>=1.8
scipy>=0.14
pandas>=0.15
biopython>=1.65
setuptools>=2.0
argparse>=1.1
BeautifulSoup4>=4.3.2
